Refactor some names on facts module
