+++ features/facts/src/main/java/io/dotanuki/norris/facts/data/ActualSearchDataSource.kt
@@ -2,11 +2,7 @@ package io.dotanuki.norris.facts.data
 
 import io.dotanuki.norris.persistance.LocalStorage
 
-class LatestSearchesDataSource(private val storage: LocalStorage) {
-
-    fun save(query: String) {
-        storage.registerNewSearch(query)
-    }
+class ActualSearchDataSource(private val storage: LocalStorage) {
 
     suspend fun actualQuery(): String =
         with(storage.lastSearches()) {
