+++ features/facts/src/test/java/io/dotanuki/norri/facts/FactsPresentationTests.kt
@@ -1,8 +1,7 @@
 package io.dotanuki.norri.facts
 
 import com.google.common.truth.Truth.assertThat
-import io.dotanuki.norris.domain.model.ChuckNorrisFact
-import io.dotanuki.norris.domain.model.RelatedCategory
+import io.dotanuki.norris.facts.domain.ChuckNorrisFact
 import io.dotanuki.norris.facts.presentation.FactDisplayRow
 import io.dotanuki.norris.facts.presentation.FactsPresentation
 import org.junit.Test
@@ -15,13 +14,11 @@ class FactsPresentationTests {
                 id = "lhan43nqsgowtaffzxouua",
                 shareableUrl = "https://api.chucknorris.io/jokes/lhan43nqsgowtaffzxouua",
                 textual = "Chuck Norris divides by zero.",
-                category = RelatedCategory.Available("math")
             ),
             ChuckNorrisFact(
                 id = "2wzginmks8azrbaxnamxdw",
                 shareableUrl = "https://api.chucknorris.io/jokes/2wzginmks8azrbaxnamxdw",
                 textual = "As 7 maravilhas do mundo moderno são: a mão direita e esquerda de Chuck Norris, seu pé direito e esquerdo, seu cinturão, seu chapéu e sua barba",
-                category = RelatedCategory.Missing
             )
         )
 
@@ -29,13 +26,11 @@ class FactsPresentationTests {
             "god like",
             listOf(
                 FactDisplayRow(
-                    tag = RelatedCategory.Available("math"),
                     url = "https://api.chucknorris.io/jokes/lhan43nqsgowtaffzxouua",
                     fact = "Chuck Norris divides by zero.",
                     displayWithSmallerFontSize = false
                 ),
                 FactDisplayRow(
-                    tag = RelatedCategory.Missing,
                     url = "https://api.chucknorris.io/jokes/2wzginmks8azrbaxnamxdw",
                     fact = "As 7 maravilhas do mundo moderno são: a mão direita e esquerda de Chuck Norris, seu pé direito e esquerdo, seu cinturão, seu chapéu e sua barba",
                     displayWithSmallerFontSize = true
